Refactoring for better overview
