+++ src/gui/Title.java
@@ -23,9 +23,9 @@ import java.awt.image.BufferedImage;
 import java.awt.event.MouseEvent;
 import java.awt.event.KeyEvent;
 import src.Window;
+import src.gpc.Header;
 import src.Engine;
 import src.GUI;
-import src.obj.Header;
 
 public class Title extends GUI implements Start {
